clean and refactor the code
