+++ tests/test_token_counter.py
@@ -1,7 +1,7 @@
 import unittest
 
 import tests.context
-from autogpt.token_counter import count_message_tokens, count_string_tokens
+from autogpt.llm_utils import count_message_tokens, count_string_tokens
 
 
 class TestTokenCounter(unittest.TestCase):
